Fuse ServiceManagers
