+++ avaje-jex/src/main/java/io/avaje/jex/core/Constants.java
@@ -16,7 +16,7 @@ public final class Constants {
   public static final String HOST = "Host";
   public static final String USER_AGENT = "User-Agent";
   public static final String ACCEPT_ENCODING = "Accept-Encoding";
-  
+
   public static final String TEXT_HTML = "text/html";
   public static final String TEXT_PLAIN = "text/plain";
   public static final String TEXT_HTML_UTF8 = "text/html;charset=utf-8";
